fix(entity-management): Abolish address referencing
Addresses with the same content were not saved as new DB entities, but as references to the same object. This meant that if, for example, the address of an imported farmer was changed, all addresses of imported farmers were changed. With the new change, a new address entity is always created and never a reference.

Resolves FOR-580